Refactor: 상태처리 App에서 TextBox로 이동
